Export the InferOptions and InferArguments type helpers
This allows callers to write helper methods that manipulate commands (eg
to add options to a preexisting command)

This is technically already possible if you rely on type inference, but
I've run into:

error TS7056: The inferred type of this node exceeds the maximum length
the compiler will serialize. An explicit type annotation is needed.